requirements.txt
asgiref==3.2.10
attrs==20.3.0
dj-database-url==0.5.0
Django==3.1.2
django-environ==0.4.5
django-heroku==0.3.1
django-rest-authtoken==2.1.3
djangorestframework==3.12.1
drf-spectacular==0.11.1
gunicorn==20.0.4
inflection==0.5.1
jsonschema==3.2.0
psycopg2==2.8.6
pyrsistent==0.17.3
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
whitenoise==5.2.0
